improvement: Reduced usage of aux functions

+++ cursus/01-ft_printf/libft/libft.h
@@ -49,7 +49,7 @@ void	ft_striteri(char *s, void (*f)(unsigned int, char*));
 int		ft_putchar_fd(char c, int fd);
 int		ft_putstr_fd(char *s, int fd);
 int		ft_putendl_fd(char *s, int fd);
-int		ft_putnbr_fd(int n, int fd);
+int		ft_putnbr_fd(long n, int fd);
 
 typedef struct s_list
 {
